+++ ext/stub.c
@@ -169,6 +169,7 @@ zend_class_entry *stub_issue2546_ce;
 zend_class_entry *stub_issue663_ce;
 zend_class_entry *stub_issue808_ce;
 zend_class_entry *stub_issue820_ce;
+zend_class_entry *stub_issue882_ce;
 zend_class_entry *stub_issue887_ce;
 zend_class_entry *stub_issue893_ce;
 zend_class_entry *stub_issue914_ce;
@@ -422,6 +423,7 @@ static PHP_MINIT_FUNCTION(stub)
 	ZEPHIR_INIT(Stub_Issue663);
 	ZEPHIR_INIT(Stub_Issue808);
 	ZEPHIR_INIT(Stub_Issue820);
+	ZEPHIR_INIT(Stub_Issue882);
 	ZEPHIR_INIT(Stub_Issue887);
 	ZEPHIR_INIT(Stub_Issue893);
 	ZEPHIR_INIT(Stub_Issue914);