+++ ext/stub.h
@@ -119,6 +119,7 @@
 #include "stub/issue663.zep.h"
 #include "stub/issue808.zep.h"
 #include "stub/issue820.zep.h"
+#include "stub/issue882.zep.h"
 #include "stub/issue887.zep.h"
 #include "stub/issue893.zep.h"
 #include "stub/issue914.zep.h"